cambio de increments a bigIncrements enla migración

+++ src/migrations/2016_10_06_222558_crear_modulo_examen_coprologico.php
@@ -13,8 +13,8 @@ class CrearModuloExamenCoprologico extends Migration {
     public function up()
     {
         Schema::create('examen_coprologico', function($table){
-            $table->increments('id');
-            $table->integer('id_consulta')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('id_consulta')->unsigned();
             $table->string('color')->nullable();
             $table->string('consistencia')->nullable();
             $table->string('flora_bacteriana')->nullable();
